Reorganize Tray Code
Remove tray code from bar.c and render.c
